Improve the idle command of Ompdc
Left margin for songs is done with I.hpad instead of spaces in string.
The interface now respond to Ctrl+c in order to quit.
When the interface is redrawn, a timestamp is checked (wait 1 second
between each request at least) in order to not make to much request to
the mpd server. For example on resize events, there can be a lot of redraw
in a little span of time that could make ompdc to hang.